* Updated version of the SOAP container by Bruno Pedro
Changelog:

The container has been changed because some webservices don't return
usernames and passwords, returning an error when the login is invalid.

The option 'matchpasswords' has been added that is set to true by default,
not breaking backwards compatibility. If 'matchpassword' is false, local
and remote passwords are not compared.


git-svn-id: http://svn.php.net/repository/pear/packages/Auth/trunk@133461 c90b9560-bf6c-de11-be94-00142212c4b1